modified code corrections according to the liking of the supreme commanda
