+++ app/users/auth_routes.py
@@ -6,7 +6,7 @@ from authlib.integrations.starlette_client import OAuth
 from fastapi.security import OAuth2PasswordRequestForm
 from ..db.database import Database
 from ..deps import get_db_connection
-from .crud import add_user_to_db, get_user_from_db, update_refresh_token_on_db, revoke_user_token_on_db
+from .authentication_crud import add_user_to_db, get_user_from_db, update_refresh_token_on_db, revoke_user_token_on_db
 from ..utils.logger import logger
 import asyncpg
 from ..config.settings import settings as st
